Ticket #641   The merge will call the events within the existing ophelia model directly as opposed to calling them via the network interface.
  Ticket #647
  Delete the draft when the publish event occurs.

+++ remote/model/src/main/java/com/thinkparity/desdemona/model/profile/ProfileModelImpl.java
@@ -252,7 +252,15 @@ public final class ProfileModelImpl extends AbstractModelImpl implements
             final Token existingToken = userSql.readProfileToken(userId);
             if (null != existingToken) {
                 getQueueModel().deleteEvents(userId);
-                getArtifactModel().deleteDrafts(userId, now);
+                /* HACK - up until a point, the "thinkParity" user was not able
+                 * to login - now this user does login such that certain apis
+                 * can be exercised with an actual user login - threfore we
+                 * need to ignore the user in this scenario */
+                if (userId.equals(User.THINKPARITY.getId())) {
+                    logger.logInfo("Logging in as system user.");
+                } else {
+                    getArtifactModel().deleteDrafts(userId, now);
+                }
             }
 
             final Token newToken = newToken();
